projecr_drf/requirements.txt
django 
djangorestframework 
djoser